@@ -1,32 +1,39 @@
 {
-  "name": "reading-list-browser-extension",
+  "name": "nsx-browser-extension",
   "version": "0.0.1",
-  "description": "ReadlingList Browser Extension",
+  "type": "module",
+  "description": "NSX Browser Extension",
   "volta": {
-    "node": "16.20.1"
+    "node": "22.20.0"
   },
   "scripts": {
+    "dev": "wxt",
+    "dev:firefox": "wxt -b firefox",
+    "build": "wxt build",
+    "build:firefox": "wxt build -b firefox",
+    "zip": "wxt zip",
+    "clean": "rm -rf .output",
     "typecheck": "tsc --noEmit",
-    "lint:fix": "eslint src --ext .ts,.tsx,.js,jsx --fix",
-    "test": "mocha",
-    "clean": "rm -rf dist",
-    "prebuild": "npm run clean",
-    "init:profile": "mkdir -p tmp/profile/chrome tmp/profile/brave tmp/profile/firefox",
-    "build": "NODE_ENV=production webpack",
-    "build:chrome": "NODE_ENV=production webpack --config ./webpack/chrome.js",
-    "build:firefox": "NODE_ENV=development webpack --config ./webpack/firefox.js",
-    "build:safari": "NODE_ENV=development webpack --config ./webpack/safari.js && xcrun safari-web-extension-converter dist/safari-prebuild --project-location dist/safari --force --no-open && rm -rf safari-prebuild",
-    "watch": "NODE_ENV=development webpack -w",
-    "watch:chrome": "NODE_ENV=development webpack --config ./webpack/chrome.js -w",
-    "watch:firefox": "NODE_ENV=development webpack --config ./webpack/firefox.js -w",
-    "watch:safari": "npm run init:safari && NODE_ENV=development webpack --config ./webpack/safari.js -w",
-    "reload:chrome": "web-ext --source-dir=dist/chrome run -t chromium --chromium-binary  /Applications/Google\\ Chrome.app/Contents/MacOS/Google\\ Chrome --start-url example.com --chromium-profile tmp/profile/chrome --profile-create-if-missing --keep-profile-changes",
-    "reload:brave": "web-ext --source-dir=dist/chrome run -t chromium --chromium-binary /Applications/Brave\\ Browser.app/Contents/MacOS/Brave\\ Browser --start-url example.com --chromium-profile tmp/profile/brave --profile-create-if-missing --keep-profile-changes",
-    "reload:firefox": "web-ext --source-dir=dist/firefox run --start-url example.com --firefox-profile tmp/profile/firefox --profile-create-if-missing --keep-profile-changes",
-    "start:chrome": "concurrently npm:watch:chrome npm:reload:chrome",
-    "start:brave": "concurrently npm:watch:chrome npm:reload:brave",
-    "start:firefox": "concurrently npm:watch:firefox npm:reload:firefox",
-    "start": "concurrently npm:watch:chrome npm:reload:brave npm:watch:firefox npm:reload:firefox"
+    "lint": "eslint src --ext .ts,.tsx,.js,.jsx",
+    "lint:fix": "eslint src --ext .ts,.tsx,.js,.jsx --fix",
+    "test": "playwright test",
+    "test:ui": "playwright test --ui",
+    "test:debug": "playwright test --debug",
+    "test:unit": "vitest",
+    "test:unit:ui": "vitest --ui",
+    "test:unit:run": "vitest run",
+    "test:coverage": "vitest run --coverage",
+    "test:all": "pnpm test:unit:run && pnpm test",
+    "prepare": "husky"
+  },
+  "lint-staged": {
+    "*.{ts,tsx,js,jsx}": [
+      "eslint --fix",
+      "prettier --write"
+    ],
+    "*.{json,md,css}": [
+      "prettier --write"
+    ]
   },
   "author": "Debdut Karmakar <iawaiponly@gmail.com>",
   "license": "ISC",
