@@ -1,3 +1,6 @@
-module.exports = {
-  plugins: [require('tailwindcss'), require('autoprefixer')],
+import tailwindcss from '@tailwindcss/postcss'
+import autoprefixer from 'autoprefixer'
+
+export default {
+  plugins: [tailwindcss, autoprefixer],
 }
