@@ -3,5 +3,7 @@ import { createRoot } from 'react-dom/client'
 
 import App from './components/app'
 
-const root = createRoot(document.querySelector('div#popup'))
+const popupElement = document.querySelector('div#popup')
+if (!popupElement) throw new Error('Popup element not found')
+const root = createRoot(popupElement)
 root.render(<App />)
